Ensure reconnects to Platform due to deployments are less noisy
The client logs on every disconnect, even if the reconnect immediately after succeeds. This makes deployments and unstable internet very noisy for our customers.

This PR brings an update to the logging, so the user, after initial connection, is only notified of loss of connection and re-attaining at the tenth retry, which means the application has not been connected for 10 seconds.

Additionally, I updated the poms and workflows for the axon-4 branch, to continue bugfixes here for AF4. After merge I will port the fix to AF5.