Ensure reconnects to Platform due to deployments are less noisy (AF5)
The client logs on every disconnect, even if the reconnect immediately after succeeds. This makes deployments and unstable internet very noisy for our customers.

This PR brings an update to the logging, so the user, after initial connection, is only notified of loss of connection and re-attaining at the tenth retry, which means the application has not been connected for 10 seconds.